Makefile Rules.mk bzip2/bzlib.c bzip2/bzlib_private.h cassette.cpp cassette.h spc1000.cpp spc1000.h ../emu2149 ../mame/backup/src/devices/bus/spc1000/miniz_export.h ../sdl2/ugui.c

+++ sdl2/ugui.c
@@ -5402,9 +5402,9 @@ void _UG_PutChar( char chr, UG_S16 x, UG_S16 y, UG_COLOR fc, UG_COLOR bc, const 
 			  for( i=0;i<actual_char_width;i++ )
 			  {
 				 b = font->p[index++];
-				 color = (((fc & 0xFF) * b + (bc & 0xFF) * (256 - b)) >> 8) & 0xFF |//Blue component
-				         (((fc & 0xFF00) * b + (bc & 0xFF00) * (256 - b)) >> 8)  & 0xFF00|//Green component
-				         (((fc & 0xFF0000) * b + (bc & 0xFF0000) * (256 - b)) >> 8) & 0xFF0000; //Red component
+				 color = (unsigned short)((((fc & 0xFF) * b + (bc & 0xFF) * (256 - b)) >> 8)         &     0xFF) |//Blue component
+                     (unsigned short)((((fc & 0xFF00) * b + (bc & 0xFF00) * (256 - b)) >> 8)     &   0xFF00) |//Green component
+				         (unsigned short)((((fc & 0xFF0000) * b + (bc & 0xFF0000) * (256 - b)) >> 8) & 0xFF0000) ; //Red component
 				 push_pixel(color);
 			  }
 			  index += font->char_width - actual_char_width;
